requirements.txt
decouple==0.0.7
Django==4.2.3
django-cors-headers==4.2.0
django-import-export==3.2.0
django-redis==5.3.0
django-timezone-field==5.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
psycopg2-binary==2.9.7
PyJWT==2.7.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8

